requirements.txt
numpy
pandas
opencv-python
tqdm
vocab
revtok
numpy
Pillow
sacred
etaprogress
scikit-video
lmdb
gtimer
filelock
networkx
termcolor
torch==1.7.1
torchvision==0.8.2
tensorboardX==1.8
ai2thor==2.1.0
